ARM32 mbedTLS 2.16.10 porting: cross-compile, struct size, CENC multi-sample, entropy, double-free fixes
- Header rename: pkcs8.h/test_helpers.h -> *_mbedtls.h/*_openssl.h to avoid shadowing
- CMake: FetchContent offline mode, -march=native switch, pthread/GTest/OpenSSL fixes
- CENC: sa_subsample_length_s struct size fix (1031 tests), offset propagation (multi-sample)
- EC ElGamal: parameter struct conversion for ARM32 size_t vs uint64_t (70 tests)
- symmetric: CTR cipher reinit from key length instead of stale context
- typej: double-free fix (NULL after free on error path)
- hardware_rng: /dev/hwrng -> /dev/urandom fallback with read validation
- random(): chunked CTR-DRBG generation for >1024 byte requests
- object_store tests: proper cleanup to eliminate leak warnings
- sa_ta_types.h: _Static_assert for sa_subsample_length_s == 16 bytes
- Compiler warning fixes for ARM32 cross-compilation

+++ reference/src/taimpl/src/porting/memory.c
@@ -77,8 +77,8 @@ bool memory_is_valid_clear(
         return false;
     }
 
-    size_t temp;
-    if (add_overflow((unsigned long) memory_location, size, &temp)) {
+    unsigned long temp;
+    if (add_overflow((unsigned long) memory_location, (unsigned long) size, &temp)) {
         ERROR("Integer overflow");
         return false;
     }
